add SCP-02 spec to Arduino simulator and test app
modified the Arduino scale simulator to allow the user interface to
choose either SCP-01 or SCP-02 scale protocol specification. Using the
letter C of the membrane keypad and then choose digit to choose the
spec: 0 for SCP-01 or 1 for SCP-02.

modified the test application to properly parse either an SCP-01 or an
SCP-02 protocol message.