Update serialcommands.ino
added LCD interface to provide a way to display settings.

Added #defines for USE_LCD, USE_KEYPAD to allow or disallow user interface.

For testing uncomment USE_SERIAL which will print messages to serial port.  This should be disabled by commenting out the define when used with an actual point of sale since the output is not compatible with the scale control protocol.

While testing, found a couple of issues in the keypad state machine that required some logic and state changes.

Changed the default start up from the SCP-01 protocol to the SCP02 protocol as that is what GenPOS point of sale uses.